backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.7.0
django-dotenv==1.4.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dj-database-url==2.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
Pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
